PowerHAL: Fix performance locking list search logic
The function get_locktype_by_type returns an error when the
specified lock cannot be found in the global lock list.
In function new_lock_init, the logic was inverted, causing
missed performance locking and complete breakage of the
Advanced Power Hints logic.

The logic breakage is due to the loss of a commit at final
phase, when a logic inversion of the search functions was
ddone.